# docs/swagger.yaml
basePath: /
definitions:
...
  postgres.Submission:
    properties:
      code:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      execution_time_ms:
        type: integer
      id:
        type: integer
      language:
        type: string
      memory_usage_mb:
        type: integer
      problem_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple online judge
  title: Adel Online Judge API
  version: "1.0"
paths:
...
  /submissions:
    post:
      consumes:
      - application/json
      description: Create a new code submission for a problem
      parameters:
      - description: Submission object
        in: body
        name: submission
        required: true
        schema:
          $ref: '#/definitions/postgres.Submission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                submission:
                  $ref: '#/definitions/postgres.Submission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create submission
      tags:
      - submissions
  /submissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a submission by ID
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete submission
      tags:
      - submissions
    get:
      consumes:
      - application/json
      description: Get a single submission by its ID
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                submission:
                  $ref: '#/definitions/postgres.Submission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get submission by ID
      tags:
      - submissions
    put:
      consumes:
      - application/json
      description: Update submission results (used by judge system)
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Submission update object
        in: body
        name: submission
        required: true
        schema:
          properties:
            error_message:
              type: string
            execution_time_ms:
              type: integer
            memory_usage_mb:
              type: integer
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                submission:
                  $ref: '#/definitions/postgres.Submission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Update submission
      tags:
      - submissions
  /submissions/user:
    get:
      consumes:
      - application/json
      description: Get all submissions for the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                submissions:
                  items:
                    $ref: '#/definitions/postgres.Submission'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get user submissions
      tags:
      - submissions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"